Refactored Image Formats management into its own class
