+++ src/ImageSharp/Configuration.cs
@@ -27,26 +27,6 @@ namespace SixLabors.ImageSharp
         private static readonly Lazy<Configuration> Lazy = new Lazy<Configuration>(CreateDefaultInstance);
 
         /// <summary>
-        /// The list of supported <see cref="IImageEncoder"/> keyed to mime types.
-        /// </summary>
-        private readonly ConcurrentDictionary<IImageFormat, IImageEncoder> mimeTypeEncoders = new ConcurrentDictionary<IImageFormat, IImageEncoder>();
-
-        /// <summary>
-        /// The list of supported <see cref="IImageEncoder"/> keyed to mime types.
-        /// </summary>
-        private readonly ConcurrentDictionary<IImageFormat, IImageDecoder> mimeTypeDecoders = new ConcurrentDictionary<IImageFormat, IImageDecoder>();
-
-        /// <summary>
-        /// The list of supported <see cref="IImageFormat"/>s.
-        /// </summary>
-        private readonly ConcurrentBag<IImageFormat> imageFormats = new ConcurrentBag<IImageFormat>();
-
-        /// <summary>
-        /// The list of supported <see cref="IImageFormatDetector"/>s.
-        /// </summary>
-        private ConcurrentBag<IImageFormatDetector> imageFormatDetectors = new ConcurrentBag<IImageFormatDetector>();
-
-        /// <summary>
         /// Initializes a new instance of the <see cref="Configuration" /> class.
         /// </summary>
         public Configuration()
@@ -81,7 +61,12 @@ namespace SixLabors.ImageSharp
         /// <summary>
         /// Gets the currently registered <see cref="IImageFormat"/>s.
         /// </summary>
-        public IEnumerable<IImageFormat> ImageFormats => this.imageFormats;
+        public IEnumerable<IImageFormat> ImageFormats => this.FormatsManager.ImageFormats;
+
+        /// <summary>
+        /// Gets or sets the <see cref="ImageFormatsManager"/> that is currently in use.
+        /// </summary>
+        public ImageFormatsManager FormatsManager { get; set; } = new ImageFormatsManager();
 
         /// <summary>
         /// Gets or sets the <see cref="MemoryManager"/> that is currently in use.
@@ -91,22 +76,22 @@ namespace SixLabors.ImageSharp
         /// <summary>
         /// Gets the maximum header size of all the formats.
         /// </summary>
-        internal int MaxHeaderSize { get; private set; }
+        internal int MaxHeaderSize => this.FormatsManager.MaxHeaderSize;
 
         /// <summary>
         /// Gets the currently registered <see cref="IImageFormatDetector"/>s.
         /// </summary>
-        internal IEnumerable<IImageFormatDetector> FormatDetectors => this.imageFormatDetectors;
+        internal IEnumerable<IImageFormatDetector> FormatDetectors => this.FormatsManager.FormatDetectors;
 
         /// <summary>
         /// Gets the currently registered <see cref="IImageDecoder"/>s.
         /// </summary>
-        internal IEnumerable<KeyValuePair<IImageFormat, IImageDecoder>> ImageDecoders => this.mimeTypeDecoders;
+        internal IEnumerable<KeyValuePair<IImageFormat, IImageDecoder>> ImageDecoders => this.FormatsManager.ImageDecoders;
 
         /// <summary>
         /// Gets the currently registered <see cref="IImageEncoder"/>s.
         /// </summary>
-        internal IEnumerable<KeyValuePair<IImageFormat, IImageEncoder>> ImageEncoders => this.mimeTypeEncoders;
+        internal IEnumerable<KeyValuePair<IImageFormat, IImageEncoder>> ImageEncoders => this.FormatsManager.ImageEncoders;
 
 #if !NETSTANDARD1_1
         /// <summary>
@@ -136,10 +121,7 @@ namespace SixLabors.ImageSharp
         /// <param name="format">The format to register as a known format.</param>
         public void AddImageFormat(IImageFormat format)
         {
-            Guard.NotNull(format, nameof(format));
-            Guard.NotNull(format.MimeTypes, nameof(format.MimeTypes));
-            Guard.NotNull(format.FileExtensions, nameof(format.FileExtensions));
-            this.imageFormats.Add(format);
+            this.FormatsManager.AddImageFormat(format);
         }
 
         /// <summary>
@@ -149,7 +131,7 @@ namespace SixLabors.ImageSharp
         /// <returns>The <see cref="IImageFormat"/> if found otherwise null</returns>
         public IImageFormat FindFormatByFileExtension(string extension)
         {
-            return this.imageFormats.FirstOrDefault(x => x.FileExtensions.Contains(extension, StringComparer.OrdinalIgnoreCase));
+            return this.FormatsManager.FindFormatByFileExtension(extension);
         }
 
         /// <summary>
@@ -159,7 +141,7 @@ namespace SixLabors.ImageSharp
         /// <returns>The <see cref="IImageFormat"/> if found; otherwise null</returns>
         public IImageFormat FindFormatByMimeType(string mimeType)
         {
-            return this.imageFormats.FirstOrDefault(x => x.MimeTypes.Contains(mimeType, StringComparer.OrdinalIgnoreCase));
+            return this.FormatsManager.FindFormatByMimeType(mimeType);
         }
 
         /// <summary>
@@ -169,10 +151,7 @@ namespace SixLabors.ImageSharp
         /// <param name="encoder">The encoder to use,</param>
         public void SetEncoder(IImageFormat imageFormat, IImageEncoder encoder)
         {
-            Guard.NotNull(imageFormat, nameof(imageFormat));
-            Guard.NotNull(encoder, nameof(encoder));
-            this.AddImageFormat(imageFormat);
-            this.mimeTypeEncoders.AddOrUpdate(imageFormat, encoder, (s, e) => encoder);
+            this.FormatsManager.SetEncoder(imageFormat, encoder);
         }
 
         /// <summary>
@@ -182,10 +161,7 @@ namespace SixLabors.ImageSharp
         /// <param name="decoder">The decoder to use,</param>
         public void SetDecoder(IImageFormat imageFormat, IImageDecoder decoder)
         {
-            Guard.NotNull(imageFormat, nameof(imageFormat));
-            Guard.NotNull(decoder, nameof(decoder));
-            this.AddImageFormat(imageFormat);
-            this.mimeTypeDecoders.AddOrUpdate(imageFormat, decoder, (s, e) => decoder);
+            this.FormatsManager.SetDecoder(imageFormat, decoder);
         }
 
         /// <summary>
@@ -193,7 +169,7 @@ namespace SixLabors.ImageSharp
         /// </summary>
         public void ClearImageFormatDetectors()
         {
-            this.imageFormatDetectors = new ConcurrentBag<IImageFormatDetector>();
+            this.FormatsManager.ClearImageFormatDetectors();
         }
 
         /// <summary>
@@ -202,9 +178,7 @@ namespace SixLabors.ImageSharp
         /// <param name="detector">The detector to add</param>
         public void AddImageFormatDetector(IImageFormatDetector detector)
         {
-            Guard.NotNull(detector, nameof(detector));
-            this.imageFormatDetectors.Add(detector);
-            this.SetMaxHeaderSize();
+            this.FormatsManager.AddImageFormatDetector(detector);
         }
 
         /// <summary>
@@ -214,13 +188,7 @@ namespace SixLabors.ImageSharp
         /// <returns>The <see cref="IImageDecoder"/> if found otherwise null</returns>
         public IImageDecoder FindDecoder(IImageFormat format)
         {
-            Guard.NotNull(format, nameof(format));
-            if (this.mimeTypeDecoders.TryGetValue(format, out IImageDecoder decoder))
-            {
-                return decoder;
-            }
-
-            return null;
+            return this.FormatsManager.FindDecoder(format);
         }
 
         /// <summary>
@@ -230,13 +198,7 @@ namespace SixLabors.ImageSharp
         /// <returns>The <see cref="IImageEncoder"/> if found otherwise null</returns>
         public IImageEncoder FindEncoder(IImageFormat format)
         {
-            Guard.NotNull(format, nameof(format));
-            if (this.mimeTypeEncoders.TryGetValue(format, out IImageEncoder encoder))
-            {
-                return encoder;
-            }
-
-            return null;
+            return this.FormatsManager.FindEncoder(format);
         }
 
         /// <summary>
@@ -254,14 +216,6 @@ namespace SixLabors.ImageSharp
                 new JpegConfigurationModule(),
                 new GifConfigurationModule(),
                 new BmpConfigurationModule());
-        }
-
-        /// <summary>
-        /// Sets the max header size.
-        /// </summary>
-        private void SetMaxHeaderSize()
-        {
-            this.MaxHeaderSize = this.imageFormatDetectors.Max(x => x.HeaderSize);
         }
     }
 }
